refactor: move misplaced functions to correct domains
Move pure AST queries and type predicates out of codegen files into
their proper modules (ast.c, types.c, print_ast.c) and eliminate
duplicate implementations across checker and codegen.

- Eliminate tuple_types_equal (codegen.c) — use type_equals() instead
- Eliminate find_resolved_enum_variant (checker_expr.c) — use type_enum_variant_index()
- Add type_find_field_index to types.c, replacing find_struct_field_index
  (checker.c) and find_struct_field_type (codegen_stmt.c)
- Move type_from_node from codegen.c to types.c
- Move 5 AST query functions + helper from codegen.c to ast.c
- Move stringify_expr family (6 functions) from codegen.c to print_ast.c
- Move has_owned_temps family (3 functions) from codegen_stmt.c to ast.c
