+++ w0/codegen_internal.h
@@ -3,17 +3,15 @@
 
 #include "codegen_emit.h"
 #include "codegen_types.h"
+#include "print_ast.h"
 
 // --- From codegen.c: generic substitution ---
 Type*           subst_lookup(CodeGen* gen, const char* name);
-char*           stringify_expr(Node* node);
 int             lookup_string_lit(CodeGen* gen, const char* value, int length);
 GenericFuncDef* lookup_generic_func_def_for_instance(CodeGen* gen, const char* base_name);
-Node* find_generic_method_func_decl(Node* ast, const char* receiver_type, const char* method_name);
-Node* find_generic_func_decl(Node* ast, const char* name);
-int   parse_method_key(const char* base_name, char* recv_out, int recv_size, char* method_out,
-                       int method_size);
-void  register_thunk(CodeGen* gen, const char* c_name, Type* func_type);
+int  parse_method_key(const char* base_name, char* recv_out, int recv_size, char* method_out,
+                      int method_size);
+void register_thunk(CodeGen* gen, const char* c_name, Type* func_type);
 
 // --- From codegen_emit.c: shared helpers ---
 void        defer_push(CodeGen* gen, Node* node);
@@ -47,7 +45,6 @@ Node* lookup_generic_template_field_type(CodeGen* gen, const char* field_name);
 
 // --- From codegen_stmt.c: statement emission ---
 void emit_block_contents(CodeGen* gen, Node* block);
-int  has_owned_temps(Node* node);
 int  hoist_owned_temps(CodeGen* gen, Node* expr);
 void cleanup_owned_temps(CodeGen* gen, int saved_count);
