+++ w0/codegen_types.h
@@ -19,14 +19,6 @@ int codegen_extract_method_bindings(NodeList* pattern_args, Type** concrete_args
                                     char*** out_params, Type*** out_args, int* out_count);
 
 // Shared helpers (defined in codegen.c)
-int   tuple_types_equal(Type* a, Type* b);
-Type* type_from_node(Node* type_node);
 char* build_mangled_name_from_generic_node(CodeGen* gen, Node* type_node);
-
-// Generic AST lookup helpers (defined in codegen.c)
-Node* find_generic_struct_decl(Node* ast, const char* name);
-Node* find_generic_enum_decl(Node* ast, const char* name);
-void  collect_generic_methods(Node* ast, const char* struct_name, Node*** methods_out,
-                              int* count_out);
 
 #endif
