+++ w0/print_ast.c
@@ -1184,3 +1184,131 @@ void print_ast_checked(Node* node, int depth) {
         break;
     }
 }
+
+// ============================================================================
+// Stringify: render AST expression to string (for assert messages)
+// ============================================================================
+
+// Append a C string into a bounded stringify buffer.
+static void stringify_append_cstr(char* buf, int buf_size, int* pos, const char* s) {
+    while (*s && *pos < buf_size - 1) {
+        buf[(*pos)++] = *s++;
+    }
+}
+
+// Append a fixed-length byte slice into a bounded stringify buffer.
+static void stringify_append_n(char* buf, int buf_size, int* pos, const char* s, int n) {
+    for (int i = 0; i < n && *pos < buf_size - 1; i++) {
+        buf[(*pos)++] = s[i];
+    }
+}
+
+// Append a single character into a bounded stringify buffer.
+static void stringify_append_char(char* buf, int buf_size, int* pos, char c) {
+    if (*pos < buf_size - 1) {
+        buf[(*pos)++] = c;
+    }
+}
+
+// Append a string literal with escaping and surrounding quotes.
+static void stringify_append_string_lit(char* buf, int buf_size, int* pos, const char* s) {
+    stringify_append_cstr(buf, buf_size, pos, "\\\"");
+    while (*s && *pos < buf_size - 1) {
+        if (*s == '"') {
+            stringify_append_cstr(buf, buf_size, pos, "\\\"");
+        } else if (*s == '\\') {
+            stringify_append_cstr(buf, buf_size, pos, "\\\\");
+        } else if (*s == '\n') {
+            stringify_append_cstr(buf, buf_size, pos, "\\n");
+        } else if (*s == '\t') {
+            stringify_append_cstr(buf, buf_size, pos, "\\t");
+        } else if (*s == '\r') {
+            stringify_append_cstr(buf, buf_size, pos, "\\r");
+        } else {
+            stringify_append_char(buf, buf_size, pos, *s);
+        }
+        s++;
+    }
+    stringify_append_cstr(buf, buf_size, pos, "\\\"");
+}
+
+// Stringify an expression AST node into a human-readable string for assert messages
+static void stringify_expr_to(Node* node, char* buf, int buf_size, int* pos) {
+    if (!node || *pos >= buf_size - 1)
+        return;
+
+    switch (node->type) {
+    case NODE_INT_LIT: {
+        char tmp[32];
+        snprintf(tmp, sizeof(tmp), "%ld", node->as.int_lit.value);
+        stringify_append_cstr(buf, buf_size, pos, tmp);
+        break;
+    }
+    case NODE_FLOAT_LIT: {
+        char tmp[32];
+        snprintf(tmp, sizeof(tmp), "%g", node->as.float_lit.value);
+        stringify_append_cstr(buf, buf_size, pos, tmp);
+        break;
+    }
+    case NODE_BOOL_LIT:
+        stringify_append_cstr(buf, buf_size, pos, node->as.bool_lit.value ? "true" : "false");
+        break;
+    case NODE_STRING_LIT:
+        stringify_append_string_lit(buf, buf_size, pos, node->as.string_lit.value);
+        break;
+    case NODE_NULL_LIT:
+        stringify_append_cstr(buf, buf_size, pos, "null");
+        break;
+    case NODE_IDENT:
+        stringify_append_n(buf, buf_size, pos, node->as.ident.name, node->as.ident.length);
+        break;
+    case NODE_BINARY: {
+        stringify_append_cstr(buf, buf_size, pos, "(");
+        stringify_expr_to(node->as.binary.left, buf, buf_size, pos);
+        stringify_append_cstr(buf, buf_size, pos, " ");
+        stringify_append_cstr(buf, buf_size, pos, token_type_symbol(node->as.binary.op));
+        stringify_append_cstr(buf, buf_size, pos, " ");
+        stringify_expr_to(node->as.binary.right, buf, buf_size, pos);
+        stringify_append_cstr(buf, buf_size, pos, ")");
+        break;
+    }
+    case NODE_UNARY:
+        stringify_append_cstr(buf, buf_size, pos, token_type_symbol(node->as.unary.op));
+        stringify_expr_to(node->as.unary.operand, buf, buf_size, pos);
+        break;
+    case NODE_CALL: {
+        stringify_expr_to(node->as.call.func, buf, buf_size, pos);
+        stringify_append_cstr(buf, buf_size, pos, "(");
+        for (int i = 0; i < node->as.call.args.count; i++) {
+            if (i > 0) {
+                stringify_append_cstr(buf, buf_size, pos, ", ");
+            }
+            stringify_expr_to(node->as.call.args.nodes[i], buf, buf_size, pos);
+        }
+        stringify_append_cstr(buf, buf_size, pos, ")");
+        break;
+    }
+    case NODE_MEMBER:
+        stringify_expr_to(node->as.member.object, buf, buf_size, pos);
+        stringify_append_cstr(buf, buf_size, pos, ".");
+        stringify_append_n(buf, buf_size, pos, node->as.member.name, node->as.member.length);
+        break;
+    case NODE_INDEX:
+        stringify_expr_to(node->as.index.object, buf, buf_size, pos);
+        stringify_append_cstr(buf, buf_size, pos, "[");
+        stringify_expr_to(node->as.index.index, buf, buf_size, pos);
+        stringify_append_cstr(buf, buf_size, pos, "]");
+        break;
+    default:
+        stringify_append_cstr(buf, buf_size, pos, "...");
+        break;
+    }
+}
+
+char* stringify_expr(Node* node) {
+    static char buf[512];
+    int         pos = 0;
+    stringify_expr_to(node, buf, sizeof(buf), &pos);
+    buf[pos] = '\0';
+    return buf;
+}
