+++ w0/print_ast.h
@@ -3,7 +3,8 @@
 
 #include "ast.h"
 
-void print_ast(Node* node, int depth);
-void print_ast_checked(Node* node, int depth);
+void  print_ast(Node* node, int depth);
+void  print_ast_checked(Node* node, int depth);
+char* stringify_expr(Node* node);
 
 #endif
